:recycle: use f64 for val_step calculation

+++ downsample_rs/src/m4/simd.rs
@@ -121,7 +121,7 @@ mod tests {
 
     #[test]
     fn test_m4_simd_with_x_correct() {
-        let x = (0..100).map(|x| x as f32).collect::<Vec<f32>>();
+        let x = (0..100).map(|x| x as i32).collect::<Vec<i32>>();
         let x = Array1::from(x);
         let arr = (0..100).map(|x| x as f32).collect::<Vec<f32>>();
         let arr = Array1::from(arr);
@@ -141,7 +141,7 @@ mod tests {
 
     #[test]
     fn test_m4_simd_with_x_parallel_correct() {
-        let x = (0..100).map(|x| x as f32).collect::<Vec<f32>>();
+        let x = (0..100).map(|x| x as i32).collect::<Vec<i32>>();
         let x = Array1::from(x);
         let arr = (0..100).map(|x| x as f32).collect::<Vec<f32>>();
         let arr = Array1::from(arr);
